+++ gserver/src/typings.d.ts
@@ -1,0 +1,13 @@
+type StepSettings = string[];
+
+type PagesSettings = {
+    [page: string]: string
+};
+
+interface Settings {
+    cucumberautocomplete: {
+        steps?: StepSettings,
+        pages?: PagesSettings,
+        syncfeatures?: boolean | string
+    }
+}